Move PresentableComment to its own file
